+++ src/components/ErrorLogsView/utils/generateAffectedRecordInfo.test.js
@@ -1,6 +1,18 @@
+import { useIntl } from 'react-intl';
+
 import '../../../../test/jest/__mock__';
+
+import { getHookExecutionResult } from '@folio/stripes-data-transfer-components/jestUtils';
+
 import { generateAffectedRecordInfo } from './generateAffectedRecordInfo';
+import { translationsProperties } from '../../../../test/helpers';
 import { errorLogs } from '../../../../test/bigtest/fixtures/errorLogs';
+
+const useGenerateAffectedRecordInfo = ({ logs }) => {
+  const intl = useIntl();
+
+  return { recordInfo: generateAffectedRecordInfo(logs, intl.formatMessage) };
+};
 
 describe('generateAffectedRecordInfo', () => {
   it('should return correct instance record info without affected records (affectedRecords field is provided but empty)', () => {
@@ -11,8 +23,9 @@ describe('generateAffectedRecordInfo', () => {
       recordType: 'INSTANCE',
       affectedRecords: [],
     };
+    const { recordInfo } = getHookExecutionResult(useGenerateAffectedRecordInfo, [{ logs }], translationsProperties);
 
-    expect(generateAffectedRecordInfo(logs)).toEqual([
+    expect(recordInfo).toEqual([
       '{',
       '"Instance UUID": "5bf370e0-8cca-4d9c-82e4-5170ab2a0a39"',
       '"Instance HRID": "inst000000000022"',
@@ -28,8 +41,9 @@ describe('generateAffectedRecordInfo', () => {
       title: 'A semantic web primer',
       recordType: 'ITEM',
     };
+    const { recordInfo } = getHookExecutionResult(useGenerateAffectedRecordInfo, [{ logs }], translationsProperties);
 
-    expect(generateAffectedRecordInfo(logs)).toEqual([
+    expect(recordInfo).toEqual([
       '{',
       '"Item UUID": "5bf370e0-8cca-4d9c-82e4-5170ab2a0a39"',
       '"Item HRID": "inst000000000022"',
@@ -50,8 +64,9 @@ describe('generateAffectedRecordInfo', () => {
         affectedRecords: [],
       }],
     };
+    const { recordInfo } = getHookExecutionResult(useGenerateAffectedRecordInfo, [{ logs }], translationsProperties);
 
-    expect(generateAffectedRecordInfo(logs)).toEqual([
+    expect(recordInfo).toEqual([
       '{',
       '"Instance UUID": "5bf370e0-8cca-4d9c-82e4-5170ab2a0a39"',
       '"Instance HRID": "inst000000000022"',
@@ -63,7 +78,9 @@ describe('generateAffectedRecordInfo', () => {
   });
 
   it('should return correct record info with 2 level nesting', () => {
-    expect(generateAffectedRecordInfo(errorLogs[2].affectedRecord)).toEqual([
+    const { recordInfo } = getHookExecutionResult(useGenerateAffectedRecordInfo, [{ logs: errorLogs[2].affectedRecord }], translationsProperties);
+
+    expect(recordInfo).toEqual([
       '{',
       '"Instance UUID": "5bf370e0-8cca-4d9c-82e4-5170ab2a0a39"',
       '"Instance HRID": "inst000000000022"',
@@ -91,14 +108,15 @@ describe('generateAffectedRecordInfo', () => {
         affectedRecords: [],
       }],
     };
+    const { recordInfo } = getHookExecutionResult(useGenerateAffectedRecordInfo, [{ logs }], translationsProperties);
 
-    expect(generateAffectedRecordInfo(logs)).toEqual([
+    expect(recordInfo).toEqual([
       '{',
       '"Holdings UUID": "5bf370e0-8cca-4d9c-82e4-5170ab2a0a39"',
       '"Holdings HRID": "inst000000000022"',
       '"Associated Instance UUID": "e3ff6133-b9a2-4d4c-a1c9-dc1867d4df19"',
       '"Associated Instance HRID": "hold000000000009"',
-      '"Instance Title": "A semantic web primer"',
+      '"Associated Instance Title": "A semantic web primer"',
       '}',
     ]);
   });